requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

pyautogen==0.2.28
python-dotenv==1.0.1
requests==2.32.3
serpapi==0.1.5
SpeechRecognition==3.10.4
